Autogenerate ChangeLog as part of "make dist"
We use the AX_GENERATE_CHANGELOG macro from
https://www.gnu.org/software/autoconf-archive/ax_generate_changelog.html
so that ChangeLog is produced as part of "make dist"

Manually updating the ChangeLog file at each commit is an error-prone,
and ancient pattern, that is no longer used in modern projects.

"make dist" will now generate in the .tar.gz/.zip an equivalently
detailed changelog with the output of "git log --name-status".