Fix libtiff version in pkg-config
This fixes a regression introduced in a28689d5d5a1fe87370ff01063d88db42d4bf8a6
which switched the VERSION variables to the implicit ones that the cmake
"project" command defines.

Prefer LIBTIFF_VERSION to PROJECT_VERSION, etc., because LIBTIFF_VERSION
is set in both CMakeLists.txt and configure.ac.

Without this, the installed libtiff-4.pc just says "Version: ".  Packages that
depend upon libtiff with a version constraint (eg: ">=4.0.0") fail to build
because they cannot determine if the installed libtiff satisfies the version
constraint.  gegl (a GIMP dependency) is an example of a widely-used package
affected by this issue.