WIP: Internal storage of rational tags changed from float to native rational values.
This MR is a clone of MR https://gitlab.com/libtiff/libtiff/-/merge_requests/235  providing incremental steps for code review and approval. After each approved step the MR will be updated with the next step. Finally, the completed MR can be merged into master branch.

The first step just inserts the basics needed for the new direct set/get interface for rationals and uses the tags XRESOLUTION and YRESOLUTION as easy example.

Second step will insert the tags XPOSITION, YPOSITION and the first array type tag REFERENCEBLACKWHITE.

The third step will start with custom tags. This takes a lot of code at once, so it might be worth cutting in several steps.

Finally, a clean-up, also for other obsolete and commented code in LibTiff, will be done.

Most of the new functions are just an adapted copy the functions already available in LibTiff.

Note: The comment "SetGetRATIONAL_directly" is a bookmark in the code to help keep track of the scattered code changes necessary for this patch. They will be removed again after the code has been checked.